extracao-clientes/requirements.txt
pycarol
pandas
sqlalchemy
mysql-connector-python
cryptography